requirements.txt
flask
gunicorn
textblob
transformers
pipeline
google.generativeai
yfinance